+++ src/AngleSharp/Dom/Geometry/IDomQuad.cs
@@ -1,43 +1,33 @@
 
-namespace AngleSharp.Dom.Geometry
+namespace AngleSharp.Dom.Geometry;
+
+/// <summary>
+/// Represents the IDomQuad interface.
+/// </summary>
+public interface IDomQuad
 {
-    using AngleSharp.Attributes;
-    using System;
+    /// <summary>
+    /// Gets the value.
+    /// </summary>
+    IDomPoint P1 { get; }
 
     /// <summary>
-    /// Represents the IDomQuad interface.
+    /// Gets the value.
     /// </summary>
-    [DomName("DOMQuad")]
-    public interface IDomQuad
-    {
-        /// <summary>
-        /// Gets the value.
-        /// </summary>
-        [DomName("p1")]
-        IDomPoint P1 { get; }
+    IDomPoint P2 { get; }
 
-        /// <summary>
-        /// Gets the value.
-        /// </summary>
-        [DomName("p2")]
-        IDomPoint P2 { get; }
+    /// <summary>
+    /// Gets the value.
+    /// </summary>
+    IDomPoint P3 { get; }
 
-        /// <summary>
-        /// Gets the value.
-        /// </summary>
-        [DomName("p3")]
-        IDomPoint P3 { get; }
+    /// <summary>
+    /// Gets the value.
+    /// </summary>
+    IDomPoint P4 { get; }
 
-        /// <summary>
-        /// Gets the value.
-        /// </summary>
-        [DomName("p4")]
-        IDomPoint P4 { get; }
-
-        /// <summary>
-        /// Executes GetBounds and returns a value.
-        /// </summary>
-        [DomName("getBounds")]
-        IDomRect GetBounds();
-    }
+    /// <summary>
+    /// Executes GetBounds and returns a value.
+    /// </summary>
+    IDomRect GetBounds();
 }